requirements.txt
absl-py==0.8.0
appnope==0.1.0
astor==0.8.0
backcall==0.1.0
Click==7.0
cycler==0.10.0
decorator==4.4.0
Flask==1.1.1
Flask-Cors==3.0.8
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
joblib==0.13.2
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.17.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
python-igraph==0.7.1.post6
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tqdm==4.35.0
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.6
wrapt==1.11.2
